refactor(agentstatus): split installer policy
