+++ services/tuttid/service/agentstatus/installer.go
@@ -361,19 +361,6 @@ func (s Service) nextMissingInstaller(spec ProviderSpec, runtime providerRuntime
 	return InstallerSpec{}, false, ""
 }
 
-func isCodexAppServerCommand(command []string) bool {
-	return len(command) >= 2 && strings.EqualFold(strings.TrimSpace(command[1]), "app-server")
-}
-
-func legacyRuntimeFailureRequiresRepair(spec ProviderSpec, reasonCode string) bool {
-	switch strings.TrimSpace(reasonCode) {
-	case "acp_adapter_launch_failed":
-		return !isCodexStatusSpec(spec)
-	default:
-		return false
-	}
-}
-
 func installNodeForTarget(target string) string {
 	if strings.TrimSpace(target) == "adapter" {
 		return "install_adapter"
